Add new ExchangeVersionType values
We have rev'd the schema in the service and need to reflect these
changes in the client API.  This is the first change.  Note that
Exchange2015 and Exchange2016 represent the same version.  We gave the
schema the name Exchange2015 before we knew what the shipped product
name would be.  Note that we have changed the format of our versioning
identifiers as can be surmised by looking at the newest value in the
enum.