Update EwsUtilities.cs
Change the Time attribute in the trace tag to format "O" from "u", so that we capture fractions of a second too.  Without this, we need to analyze the headers of the responses to get this information.  This is important when tracing applications that can generate hundreds of requests in the same second.